requirements.txt
aiogram
mysql-connector-python
python-dotenv